refactor: context와 provider 분리
