+++ src/@lib/contexts/index.ts
@@ -1,3 +1,3 @@
-export * from "./NotificationContext";
-export * from "./ThemeContext";
-export * from "./UserContext";
+export * from "./notification-context";
+export * from "./theme-cntext";
+export * from "./user-context";
