+++ src/@lib/contexts/notification-context.ts
@@ -1,0 +1,21 @@
+import { createContext, useContext } from "react";
+import { Notification } from "../../types";
+
+interface NotificationContextType {
+  notifications: Notification[];
+  addNotification: (message: string, type: Notification["type"]) => void;
+  removeNotification: (id: number) => void;
+}
+
+export const NotificationContext = createContext<
+  NotificationContextType | undefined
+>(undefined);
+
+export const useNotificationContext = () => {
+  const ctx = useContext(NotificationContext);
+  if (!ctx)
+    throw new Error(
+      "useNotificationContext must be used within NotificationProvider",
+    );
+  return ctx;
+};
