+++ src/@lib/contexts/theme-cntext.ts
@@ -1,0 +1,17 @@
+import { createContext, useContext } from "react";
+
+interface ThemeContextType {
+  theme: string;
+  toggleTheme: () => void;
+}
+
+export const ThemeContext = createContext<ThemeContextType | undefined>(
+  undefined,
+);
+
+export const useThemeContext = () => {
+  const ctx = useContext(ThemeContext);
+  if (!ctx)
+    throw new Error("useThemeContext must be used within ThemeProvider");
+  return ctx;
+};
