+++ src/@lib/contexts/user-context.tsx
@@ -1,0 +1,18 @@
+import { createContext, useContext } from "react";
+import { User } from "../../types";
+
+interface UserContextType {
+  user: User | null;
+  login: (email: string, password: string) => void;
+  logout: () => void;
+}
+
+export const UserContext = createContext<UserContextType | undefined>(
+  undefined,
+);
+
+export const useUserContext = () => {
+  const ctx = useContext(UserContext);
+  if (!ctx) throw new Error("useUserContext must be used within UserProvider");
+  return ctx;
+};
