+++ src/@lib/providers/index.tsx
@@ -1,4 +1,6 @@
-import { NotificationProvider, ThemeProvider, UserProvider } from "../contexts";
+import { NotificationProvider } from "./notification-provider";
+import { ThemeProvider } from "./theme-provider";
+import { UserProvider } from "./user-provider";
 
 const Providers: React.FC<{ children: React.ReactNode }> = ({ children }) => {
   return (
