+++ src/types/index.ts
@@ -17,15 +17,3 @@ export interface Notification {
   message: string;
   type: "info" | "success" | "warning" | "error";
 }
-
-// AppContext 타입 정의
-export interface AppContextType {
-  theme: string;
-  toggleTheme: () => void;
-  user: User | null;
-  login: (email: string, password: string) => void;
-  logout: () => void;
-  notifications: Notification[];
-  addNotification: (message: string, type: Notification["type"]) => void;
-  removeNotification: (id: number) => void;
-}
