+++ vite.config.ts
@@ -1,4 +1,5 @@
 import react from "@vitejs/plugin-react-swc";
+import path from "path";
 import { defineConfig } from "vite";
 import { defineConfig as defineTestConfig, mergeConfig } from "vitest/config";
 const isCI = process.env.CI === "true";
@@ -7,6 +8,7 @@ export default mergeConfig(
   defineConfig({
     base: isCI ? "" : "/front_5th_chapter1-3/",
     plugins: [react()],
+    resolve: { alias: { "@": path.resolve(__dirname, "src") } },
   }),
   defineTestConfig({
     test: {
